web-api/pyproject.toml
[project]
name = "web-api"
version = "0.1.0"
description = "mishmish.ai backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.34.0",
    "pipecat-ai[openai,elevenlabs,deepgram,websocket,silero]>=0.0.100",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "openai-agents>=0.4.1",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "supabase>=2.24.0",
    "debugpy>=1.8.0",
    "braintrust>=0.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-watch>=4.2.0",
]

[project.scripts]
dev = "python -m debugpy --listen 0.0.0.0:5678 -m uvicorn main:app --reload --host 0.0.0.0 --port 8000"
test = "pytest"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
